Extract subscription registry into host/subscriptions.ts
Move handleSub/handleRtdbSub/handleUnsub, the session-bound sub registry
(#754) with resubscribeSessionSubs, registerListener, and ensurePortSubs into
host/subscriptions.ts. cleanupPort drops session-bound records through the
exported dropPortSessionSubs. Bodies verbatim; host.ts now holds only the
dispatch orchestrator (handleOp/handleMessage/dispatchMessage/handleTool/
cleanupPort) plus the barrel re-exports. No behavior change.
